move ts file from pages to component
